main/requirements.txt
requests
boto3
pillow
pymysql
